remove use of companyUpdatesEnabled flag (#641)

+++ frontend/trpc/routes/companyUpdates/index.ts
@@ -4,7 +4,7 @@ import { createInsertSchema } from "drizzle-zod";
 import { pick, truncate } from "lodash-es";
 import { z } from "zod";
 import { db } from "@/db";
-import { companyUpdates } from "@/db/schema";
+import { companyInvestors, companyUpdates } from "@/db/schema";
 import { inngest } from "@/inngest/client";
 import { type CompanyContext, companyProcedure, createRouter, renderTiptapToText } from "@/trpc";
 import { isActive } from "@/trpc/routes/contractors";
@@ -18,12 +18,18 @@ const dataSchema = createInsertSchema(companyUpdates).pick({
   body: true,
   videoUrl: true,
 });
+
+const checkHasInvestors = async (companyId: bigint) => {
+  const hasInvestors = await db.query.companyInvestors.findFirst({
+    where: eq(companyInvestors.companyId, companyId),
+  });
+  return !!hasInvestors;
+};
+
 export const companyUpdatesRouter = createRouter({
   list: companyProcedure.query(async ({ ctx }) => {
-    if (
-      !ctx.company.companyUpdatesEnabled ||
-      (!ctx.companyAdministrator && !isActive(ctx.companyContractor) && !ctx.companyInvestor)
-    )
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || (!ctx.companyAdministrator && !isActive(ctx.companyContractor) && !ctx.companyInvestor))
       throw new TRPCError({ code: "FORBIDDEN" });
     const where = and(
       eq(companyUpdates.companyId, ctx.company.id),
@@ -41,10 +47,8 @@ export const companyUpdatesRouter = createRouter({
     return { updates };
   }),
   get: companyProcedure.input(z.object({ id: z.string() })).query(async ({ ctx, input }) => {
-    if (
-      !ctx.company.companyUpdatesEnabled ||
-      (!ctx.companyAdministrator && !isActive(ctx.companyContractor) && !ctx.companyInvestor)
-    )
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || (!ctx.companyAdministrator && !isActive(ctx.companyContractor) && !ctx.companyInvestor))
       throw new TRPCError({ code: "FORBIDDEN" });
     const update = await db.query.companyUpdates.findFirst({ where: byId(ctx, input.id) });
     if (!update) throw new TRPCError({ code: "NOT_FOUND" });
@@ -56,7 +60,8 @@ export const companyUpdatesRouter = createRouter({
     };
   }),
   create: companyProcedure.input(dataSchema.required()).mutation(async ({ ctx, input }) => {
-    if (!ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || !ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
 
     const [update] = await db
       .insert(companyUpdates)
@@ -72,7 +77,8 @@ export const companyUpdatesRouter = createRouter({
     return assertDefined(update).externalId;
   }),
   update: companyProcedure.input(dataSchema.extend({ id: z.string() })).mutation(async ({ ctx, input }) => {
-    if (!ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || !ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
     const [update] = await db
       .update(companyUpdates)
       .set({
@@ -88,7 +94,8 @@ export const companyUpdatesRouter = createRouter({
     if (!update) throw new TRPCError({ code: "NOT_FOUND" });
   }),
   publish: companyProcedure.input(z.object({ id: z.string() })).mutation(async ({ ctx, input }) => {
-    if (!ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || !ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
 
     const [update] = await db
       .update(companyUpdates)
@@ -108,7 +115,8 @@ export const companyUpdatesRouter = createRouter({
     return update.externalId;
   }),
   sendTestEmail: companyProcedure.input(z.object({ id: z.string() })).mutation(async ({ ctx, input }) => {
-    if (!ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || !ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
     const update = await db.query.companyUpdates.findFirst({ where: byId(ctx, input.id) });
     if (!update) throw new TRPCError({ code: "NOT_FOUND" });
     await inngest.send({
@@ -120,7 +128,8 @@ export const companyUpdatesRouter = createRouter({
     });
   }),
   delete: companyProcedure.input(z.object({ id: z.string() })).mutation(async ({ ctx, input }) => {
-    if (!ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
+    const hasInvestors = await checkHasInvestors(ctx.company.id);
+    if (!hasInvestors || !ctx.companyAdministrator) throw new TRPCError({ code: "FORBIDDEN" });
     const result = await db.delete(companyUpdates).where(byId(ctx, input.id)).returning();
     if (result.length === 0) throw new TRPCError({ code: "NOT_FOUND" });
   }),
